+++ src/engine/QueryExecutionTree.h
@@ -11,7 +11,7 @@
 
 #include "../util/Conversions.h"
 #include "../util/HashSet.h"
-#include "../util/HttpServer/streamable_body.h"
+#include "../util/streamable_generator.h"
 #include "./Operation.h"
 #include "./QueryExecutionContext.h"
 
@@ -91,7 +91,7 @@ class QueryExecutionTree {
                            size_t limit = MAX_NOF_ROWS_IN_RESULT,
                            size_t offset = 0, char sep = '\t') const;
 
-  http_streams::stream_generator generateResults(
+  ad_utility::stream_generator::stream_generator generateResults(
       const vector<string>& selectVars, size_t limit = MAX_NOF_ROWS_IN_RESULT,
       size_t offset = 0, char sep = '\t') const;
 
@@ -205,7 +205,7 @@ class QueryExecutionTree {
       const vector<std::optional<pair<size_t, ResultTable::ResultType>>>&
           validIndices) const;
 
-  http_streams::stream_generator writeTable(
+  ad_utility::stream_generator::stream_generator writeTable(
       const IdTable& data, char sep, size_t from, size_t upperBound,
       const vector<std::optional<pair<size_t, ResultTable::ResultType>>>
           validIndices) const;
